Add blocks to multitemplate, by extending the yield system
Blocks are implemented using a mind-bending system involving shared
contexts between multiple template executions. It then pushes the
content back along the execution chain until we find a template that's
not inheriting. I need to add the super_block thing, shouldn't be too
hard, basically create a request content map that holds in progress
blocks that need their super executions. Also it's 3:30 in the morning
and I need to sleep.