Change to parsing complicated code using a mule template
So we used to try to parse complicated executable code in templates
using a complicated and buggy implementation in process_code, which was
not fully compliant with the way html/template parses the code. The
reason we weren't using html/template to parse the complicated code was
that we weren't pushing functions down into the code at the time. The
multitemplate code is now pushing functions down into template parsers,
so we can use html/template to parse compicated code, which fixes at
two bugs that I know of, and more that I don't know about. Rejoice,
0.1.beta is almost here.