More documentation, don't export extensions, rename Content to Blocks
While writing documentation, I realized Content was a bad name for the
content of blocks, so I renamed it to Blocks. Also Extensions wasn't
really intended for public use, so it was removed from the Public API.
Also wrote a sort of dictionary for the concepts of Yields and Blocks,
and cleaned up some minor documentation wordings.