Fix a bham bug, add more documentation
Realized that we could miss attributes with embedded executable lines in
them if the attribute didn't start with the executable line, so I added
a bit to consume the non-executable bit to fix that. Also updated and
moved the docuementation of the bham syntax to the doc.go file from the
README.md file.