Figured out how tokens should be organized and compiled
Big difference is that a token is now defined to have either Content or
it will have a Type and Opening, Closing and/or Children arrays that are
compiled into parse.Nodes. When there was a mix of the two, I would have
to write a bunch of extra code to then figure out the correct
interpretation.