Fix slice out of bounds, bad docs, add may_yield function
Yet another in the ongoing 0.4.X tweak-level releases, the grand change
is that we add a function. The function should have been in there
previously, but was missed. Also fix a documentation error in helpers
where we were using an older form of the struct. Finally, we can't
guarantee that prefix in the AssetLink will not be empty string, so add
a guard bit of code to fix that.